requirements.txt
absl-py==0.12.0
anyio==3.1.0
argon2-cffi==20.1.0
async-generator==1.10
atari-py==0.2.9
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
bleach==3.3.0
cached-property==1.5.2
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cirq==0.11.0
cirq-core==0.11.0
cirq-google==0.11.0
click==8.0.1
cloudpickle==1.3.0
colorama==0.4.4
configparser==5.0.2
cycler==0.10.0
decorator==4.4.2
defusedxml==0.7.1
docker-pycreds==0.4.0
entrypoints==0.3
future==0.18.2
gitdb==4.0.7
GitPython==3.1.18
google-api-core==1.31.0
google-auth==1.30.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
grpcio==1.38.0
gym==0.17.2
gym-maze-trustycoder83==0.0.4
gym-super-mario-bros==7.3.2
h5py==3.2.1
hdf5==0.1.0
idna==2.10
imageio==2.9.0
importlib-metadata==4.0.1
ipykernel==5.5.5
ipython==7.23.1
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
json5==0.9.5
jsonschema==3.2.0
kiwisolver==1.3.1
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mistune==0.8.4
mpmath==1.2.1
nbclassic==0.2.8
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nes-py==8.1.6
nest-asyncio==1.5.1
networkx==2.5.1
notebook==6.4.0
numpy==1.21.1
oauthlib==3.1.0
opencv-python==4.5.2.52
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
parso==0.8.2
pathtools==0.1.2
pickleshare==0.7.5
Pillow==8.2.0
prometheus-client==0.10.1
promise==2.3
prompt-toolkit==3.0.18
protobuf==3.13.0
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybullet==3.1.7
pycparser==2.20
pygame==1.9.6
pyglet==1.5.0
Pygments==2.9.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
pywin32==300
pywinpty==1.1.1
PyYAML==5.4.1
pyzmq==22.0.3
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.1
scikit-learn==0.24.2
scipy==1.6.3
seaborn==0.11.1
Send2Trash==1.5.0
sentry-sdk==1.1.0
shortuuid==1.0.1
six==1.16.0
smmap==4.0.0
sniffio==1.2.0
sortedcontainers==2.4.0
stable-baselines3==1.0
subprocess32==3.5.4
sympy==1.8
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.2
terminado==0.10.0
testpath==0.5.0
threadpoolctl==2.1.0
tifffile==2021.6.14
torch==1.8.1+cu111
torchaudio==0.8.1
torchvision==0.9.1+cu111
tornado==6.1
tqdm==4.61.1
traitlets==5.0.5
typing-extensions==3.10.0.0
urllib3==1.25.11
wandb==0.10.32
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.0.0
Werkzeug==2.0.1
wincertstore==0.2
zipp==3.4.1
